refactor: extract useDeckTransition hook, clean up CogTransitionExample imports
